Add L3 forwarding of mtu messages
Instead of broadcasting mtu messages in the L2 domain, enable L3
forwarding to a list of peers specified on the command line (--peers).
If using a peer_list open a raw socket to delegate routing to the OS.

Introduce additional check if the packet is sourced from a peer in
peer_list, before transmitting. Also pluck the TTL from the inbound
packet in order to decrease it before forwarding.